app/gen/requirements.txt
pandas==2.1.4
azure-storage-blob==12.19.0
google-cloud-storage==2.18.2
requests==2.31.0
load-dotenv==0.1.0
Faker==20.1.0
typer[all]==0.9.0
pytest==7.4.3
pycpfcnpj==1.7.0
